more work on enrollment status refactor
